Placed collections into their own files
